+++ Core/src/music/SyncChannel.h
@@ -8,7 +8,7 @@ public:
     SyncChannel(int numNotes, int minNote, int channel);
     ~SyncChannel();
     void pushNote(int absoluteNote, int time);
-    void tick(int currentTime);
+    void tick(long long currentTime);
     float getTimeToNext(int absoluteNote) const;
     float getTimeSinceLast(int absoluteNote) const;
     int getTotalHitsPerNote(int absoluteNote) const;
@@ -18,7 +18,7 @@ public:
 
 private:
     int _minNote;
-    int _currentTime;
+    long long _currentTime;
     std::vector<int> _lastTimePerNote;
     std::vector<std::queue<int>> _timesPerNote;
     std::vector<int> _totalHitsPerNote;
